First update that "works"
- at the fork, this project didn't do what I wanted and I had one goal:

 Build a multi-node OpenStack deployment using Vagrant

I chose to use the Liberty release.  My firt task was to get a
compantible Vagrant "box" that had the necessary software and kernel
boot parameters to keep the old "ethN" convention of network interface
naing (aka I disabled "Predictable Interface Naming"). The resulting
file size was 1GB so it won't go on github.

TODO: throw the box image in an AWS S3 bucket or something.

- I had to tune the networking a bit to get Neutron a router:external
  provider network, which is now the Vagrant eth0 interface on the
  "network" node

The originl Vagrantfile and set of scripts didn't have an externally
routable provider network.  I'm not sure if the original project ever
was able to access any of the running instances without vnc.

- Only Vagrantfile and scripts 01 and 02 have been updated

Presumably scripts 03, 04 and 05 all depend on only script 02's
successful completion (resulting in a fucntional  OpenStack deployment).
I have yet to test these.  The difficult piece was getting the Vagrantfile
and script 01 doing what was necessary to get script 02 to run to completion.
The bigger problem was that using Horizon caused the controller to go OOM just
to view a few pages.